training_pipeline/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alembic==1.13.3
    # via
    #   dagster
    #   optuna
annotated-types==0.7.0
    # via pydantic
argon2-cffi==23.1.0
    # via minio
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
cachetools==5.5.0
    # via
    #   google-auth
    #   mlflow-skinny
certifi==2024.8.30
    # via
    #   minio
    #   requests
cffi==1.17.1
    # via argon2-cffi-bindings
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   dagster
    #   mlflow-skinny
cloudpickle==3.1.0
    # via mlflow-skinny
coloredlogs==14.0
    # via dagster
colorlog==6.9.0
    # via optuna
contourpy==1.3.0
    # via matplotlib
croniter==5.0.1
    # via dagster
cycler==0.12.1
    # via matplotlib
dagster==1.8.13
    # via
    #   -r requirements.in
    #   dagster-docker
    #   dagster-postgres
dagster-docker==0.24.13
    # via -r requirements.in
dagster-pipes==1.8.13
    # via dagster
dagster-postgres==0.24.13
    # via -r requirements.in
databricks-sdk==0.36.0
    # via mlflow-skinny
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
docker==7.1.0
    # via dagster-docker
docker-image-py==0.1.13
    # via dagster-docker
docstring-parser==0.16
    # via dagster
filelock==3.16.1
    # via dagster
fonttools==4.54.1
    # via matplotlib
fsspec==2024.10.0
    # via universal-pathlib
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via mlflow-skinny
google-auth==2.35.0
    # via databricks-sdk
greenlet==3.1.1
    # via sqlalchemy
grpcio==1.67.1
    # via
    #   dagster
    #   grpcio-health-checking
grpcio-health-checking==1.62.3
    # via dagster
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via requests
importlib-metadata==8.4.0
    # via
    #   mlflow-skinny
    #   opentelemetry-api
jinja2==3.1.4
    # via dagster
joblib==1.4.2
    # via
    #   -r requirements.in
    #   scikit-learn
kiwisolver==1.4.7
    # via matplotlib
mako==1.3.6
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
matplotlib==3.9.2
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
minio==7.2.10
    # via -r requirements.in
mlflow-skinny==2.17.2
    # via -r requirements.in
numpy==2.1.2
    # via
    #   contourpy
    #   matplotlib
    #   optuna
    #   pandas
    #   scikit-learn
    #   scipy
opentelemetry-api==1.27.0
    # via
    #   mlflow-skinny
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.27.0
    # via mlflow-skinny
opentelemetry-semantic-conventions==0.48b0
    # via opentelemetry-sdk
optuna==4.0.0
    # via -r requirements.in
packaging==24.1
    # via
    #   dagster
    #   matplotlib
    #   mlflow-skinny
    #   optuna
pandas==2.2.3
    # via -r requirements.in
pillow==11.0.0
    # via matplotlib
protobuf==4.25.5
    # via
    #   dagster
    #   grpcio-health-checking
    #   mlflow-skinny
psycopg2-binary==2.9.10
    # via dagster-postgres
pyarrow==18.0.0
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pycparser==2.22
    # via cffi
pycryptodome==3.21.0
    # via minio
pydantic==2.9.2
    # via dagster
pydantic-core==2.23.4
    # via pydantic
pygments==2.18.0
    # via rich
pyparsing==3.2.0
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   croniter
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via dagster
pytz==2024.2
    # via
    #   croniter
    #   dagster
    #   pandas
pyyaml==6.0.2
    # via
    #   dagster
    #   mlflow-skinny
    #   optuna
regex==2024.9.11
    # via docker-image-py
requests==2.32.3
    # via
    #   dagster
    #   databricks-sdk
    #   docker
    #   mlflow-skinny
rich==13.9.3
    # via dagster
rsa==4.9
    # via google-auth
scikit-learn==1.5.2
    # via -r requirements.in
scipy==1.14.1
    # via
    #   -r requirements.in
    #   scikit-learn
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sqlalchemy==2.0.36
    # via
    #   alembic
    #   dagster
    #   optuna
sqlparse==0.5.1
    # via mlflow-skinny
structlog==24.4.0
    # via dagster
tabulate==0.9.0
    # via dagster
threadpoolctl==3.5.0
    # via scikit-learn
tomli==2.0.2
    # via dagster
toposort==1.10
    # via dagster
tqdm==4.66.6
    # via
    #   dagster
    #   optuna
typing-extensions==4.12.2
    # via
    #   alembic
    #   dagster
    #   minio
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzdata==2024.2
    # via pandas
universal-pathlib==0.2.5
    # via dagster
urllib3==2.2.3
    # via
    #   docker
    #   minio
    #   requests
watchdog==5.0.3
    # via dagster
wrapt==1.16.0
    # via deprecated
zipp==3.20.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
